New config file, can close running window

+++ src/circle.h
@@ -5,15 +5,15 @@
 #include <SDL2/SDL.h>
 
 typedef struct Circle {
-    int pos_x;
-    int pos_y;
-    int vel_x;
-    int vel_y;
-    int radius;
+    float pos_x;
+    float pos_y;
+    float vel_x;
+    float vel_y;
+    float radius;
     color_t *color;
 } circle_t;
 
-int circle_ctor (circle_t *c, int pos_x_, int pos_y_, int radius_, color_t *color_);
+int circle_ctor (circle_t *c, float pos_x_, float pos_y_, float radius_, color_t *color_);
 void circle_draw (SDL_Renderer *r, circle_t *c);
 void circle_move (circle_t *c);
